refactor: move navigation items to separated folder
