+++ src/components/header/header.tsx
@@ -1,6 +1,6 @@
 import { Container, Logo } from './styled';
 
-import NavigationItems from './navigation-items';
+import NavigationItems from './navigation-items/navigation-items';
 import OptionsMenu from './options-menu';
 import React from 'react';
