+++ src/components/header/navigation-items/styled.ts
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const Items = styled.ul`
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  align-content: center;
+  margin-top: 1em;
+  margin-bottom: 1em;
+  color: white;
+  user-select: none;
+  cursor: default;
+`;
